Added a stock for a moneycheck.
CheckClientMoney(client, cmoney, gunprice) will run a stock that does the math to make sure a client has enough money (true) or if he falls short (false). Also sends a message to the player if they are short. Possible to enhance this farther to show how short they are, but I had to go to work.